\name{treedata}
\docType{data}
\alias{treedata}

\title{
Tree structure Data
}

\description{
This is the tree structure data generated for Gene Ontology biological process with hgu95av2.db package version 3.2.3. Arguments are described in help file of \link{buildTree}.
}

\usage{treedata}
\format{A list contain nDAG, nTree, parentIndex, childrenIndexMat, childrenCount, componMappingMat, leaf, nonleaf, formula, sepLocation, geneSet.}
